// src/models/NotificationType.ts
export enum NotificationType {
    WARNING = "WARNING",
    AD_APPROVED = "AD_APPROVED",
    AD_REJECTED = "AD_REJECTED",
    AD_COMPLETED = "AD_COMPLETED",
    PROMOTION_APPROVED = "PROMOTION_APPROVED",
    PROMOTION_DENIED = "PROMOTION_DENIED",
    PROMOTION_CANCELED = "PROMOTION_CANCELED",
    ROOM_PROMOTION_APPROVED = "ROOM_PROMOTION_APPROVED",
    ROOM_PROMOTION_DENIED = "ROOM_PROMOTION_DENIED",
    ROOM_PROMOTION_CANCELED = "ROOM_PROMOTION_CANCELED",
    MODERATOR_ACCEPTED = "MODERATOR_ACCEPTED",
    COMMUNITY_NIGHT_REMINDER = "COMMUNITY_NIGHT_REMINDER",
    // future: SERVER_ANNOUNCEMENT, REPORT_CASE_RESOLVED
}